ui: Move lib/types.ts to lib/ui/types.ts
Change-type: patch
