+++ lib/ui/components/CardActions.tsx
@@ -7,7 +7,6 @@ import { bindActionCreators } from 'redux';
 import { Flex, Modal } from 'rendition';
 import { Form } from 'rendition/dist/unstable';
 import * as skhema from 'skhema';
-import { Card, Type } from '../../types';
 import { CardCreator } from '../components/CardCreator';
 import { FreeFieldForm } from '../components/FreeFieldForm';
 import { LINKS } from '../constants';
@@ -16,6 +15,7 @@ import { actionCreators, selectors, StoreState } from '../core/store';
 import { getLocalSchema, removeUndefinedArrayItems } from '../services/helpers';
 import { createLink } from '../services/link';
 import { createPermaLink } from '../services/url-manager';
+import { Card, Type } from '../types';
 import { ActionLink } from './ActionLink';
 import { CardLinker } from './CardLinker';
 import { ContextMenu } from './ContextMenu';
