+++ lib/ui/core/reducers/views.ts
@@ -3,9 +3,9 @@ import { JSONSchema6 } from 'json-schema';
 import * as _ from 'lodash';
 import * as skhema from 'skhema';
 import { analytics } from '../';
-import { Card } from '../../../types';
 import { hashCode } from '../../services/helpers';
 import { loadSchema } from '../../services/sdk-helpers';
+import { Card } from '../../types';
 import { Action, JellyThunkSync } from '../common';
 import { sdk } from '../sdk';
 import { actionCreators as allActionCreators, selectors, StoreState } from '../store';
