+++ lib/ui/lens/Org.tsx
@@ -15,7 +15,6 @@ import {
 import { Markdown } from 'rendition/dist/extra/Markdown';
 import { Mermaid } from 'rendition/dist/extra/Mermaid';
 import styled from 'styled-components';
-import { Card, Lens, RendererProps, Type } from '../../types';
 import { CardActions } from '../components/CardActions';
 import { CloseButton } from '../components/CloseButton';
 import Label from '../components/Label';
@@ -28,6 +27,7 @@ import {
 	getLocalSchema,
 } from '../services/helpers';
 import { createLink } from '../services/link';
+import { Card, Lens, RendererProps, Type } from '../types';
 import TimelineLens from './Timeline';
 
 const Column = styled(Flex)`
