+++ lib/ui/lens/Thread.tsx
@@ -12,7 +12,6 @@ import {
 import { Markdown } from 'rendition/dist/extra/Markdown';
 import { Mermaid } from 'rendition/dist/extra/Mermaid';
 import styled from 'styled-components';
-import { Card, Lens, RendererProps, Type } from '../../types';
 import { CardActions } from '../components/CardActions';
 import { CloseButton } from '../components/CloseButton';
 import Label from '../components/Label';
@@ -24,6 +23,7 @@ import {
 	getLocalSchema,
 } from '../services/helpers';
 import { getActor } from '../services/store-helpers';
+import { Card, Lens, RendererProps, Type } from '../types';
 import TimelineLens from './Timeline';
 
 const Column = styled(Flex)`
