+++ lib/ui/lens/support/SupportIssueTable.tsx
@@ -11,7 +11,6 @@ import {
 	Txt,
 } from 'rendition';
 import styled from 'styled-components';
-import { Card, Lens, RendererProps, Type } from '../../../types';
 import { CardCreator } from '../../components/CardCreator';
 import Icon from '../../components/Icon';
 import { actionCreators } from '../../core/store';
@@ -21,6 +20,7 @@ import {
 	getUpdateObjectFromSchema,
 	getViewSchema,
 } from '../../services/helpers';
+import { Card, Lens, RendererProps, Type } from '../../types';
 
 const Column = styled(Flex)`
 	height: 100%;
@@ -40,7 +40,7 @@ const COLUMNS = [
 		field: 'category',
 		label: 'Category',
 		sortable: true,
-	}
+	},
 	{
 		field: 'Created',
 		sortable: true,
