+++ lib/ui/lens/support/SupportThreads.tsx
@@ -10,7 +10,6 @@ import {
 	Txt,
 } from 'rendition';
 import styled from 'styled-components';
-import { Card, Lens, RendererProps } from '../../../types';
 import Gravatar from '../../components/Gravatar';
 import { actionCreators, selectors, StoreState } from '../../core/store';
 import {
@@ -20,6 +19,7 @@ import {
 	timeAgo,
 } from '../../services/helpers';
 import { getActor } from '../../services/store-helpers';
+import { Card, Lens, RendererProps } from '../../types';
 
 const Column = styled(Flex)`
 	height: 100%;
