+++ lib/ui/services/helpers.ts
@@ -5,7 +5,7 @@ import * as moment from 'moment';
 import { SchemaSieve } from 'rendition';
 import * as skhema from 'skhema';
 import uuid = require('uuid/v4');
-import { Card, Channel } from '../../types';
+import { Card, Channel } from '../types';
 
 const PURPLE = '#8268c5';
